+++ kodecks/src/shard.rs
@@ -5,14 +5,14 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Default, Serialize, Deserialize, Encode, Decode)]
 #[serde(transparent)]
-pub struct ShardList(Vec<(Color, u32)>);
+pub struct ShardList(Vec<(Color, u8)>);
 
 impl ShardList {
     pub fn new() -> Self {
         Self(Vec::new())
     }
 
-    pub fn get(&self, color: Color) -> u32 {
+    pub fn get(&self, color: Color) -> u8 {
         self.0
             .iter()
             .find(|(c, _)| *c == color)
@@ -20,7 +20,7 @@ impl ShardList {
             .unwrap_or(0)
     }
 
-    pub fn add(&mut self, color: Color, amount: u32) {
+    pub fn add(&mut self, color: Color, amount: u8) {
         self.0
             .iter_mut()
             .find(|(c, _)| *c == color)
@@ -28,7 +28,7 @@ impl ShardList {
             .unwrap_or_else(|| self.0.push((color, amount)));
     }
 
-    pub fn consume(&mut self, color: Color, amount: u32) -> Result<(), ActionError> {
+    pub fn consume(&mut self, color: Color, amount: u8) -> Result<(), ActionError> {
         let mut insufficient = true;
         if let Some((_, current)) = self.0.iter_mut().find(|(c, _)| *c == color) {
             if *current >= amount {
@@ -47,7 +47,7 @@ impl ShardList {
         }
     }
 
-    pub fn iter(&self) -> impl Iterator<Item = (Color, u32)> + '_ {
+    pub fn iter(&self) -> impl Iterator<Item = (Color, u8)> + '_ {
         self.0.iter().copied()
     }
